First commit
Example of a table that displays content from a tab-delimited resource.
The table automatically updates the number of columns as well as rows
as it reads from the resource.

To use, type in the URL of a tab delimited file or other resource in
the text field at the top, and press Enter. To demonstrate editing,
change a value in the first row and press the "Show first row" button,
which will show the contents of the first row in the console.